+++ lib/telegrammer/data_types/inline_query_result/inline_query_result_article.rb
@@ -5,9 +5,8 @@ module Telegrammer
     # @attr [String] type Type of the result, must be article
     # @attr [String] id Unique identifier for this result, 1-64 Bytes
     # @attr [String] title Title of the result
-    # @attr [String] message_text Text of the message to be sent
-    # @attr [String] parse_mode Optional. Send “Markdown”, if you want Telegram apps to show bold, italic and inline URLs in your bot's message.
-    # @attr [Boolean] disable_web_page_preview Optional. Disables link previews for links in the sent message
+    # @attr [Telegrammer::DataTypes::InputMessageContent] input_message_content Content of the message to be sent
+    # @attr [Telegrammer::DataTypes::InlineKeyboardMarkup] reply_markup Optional. Inline keyboard attached to the message
     # @attr [String] url Optional. URL of the result
     # @attr [Boolean] hide_url Optional. Pass True, if you don't want the URL to be shown in the message
     # @attr [String] description Optional. Short description of the result
@@ -20,14 +19,14 @@ module Telegrammer
       attribute :type, String, default: 'article'
       attribute :id, String
       attribute :title, String
-      attribute :message_text, String
-      attribute :parse_mode, String
-      attribute :disable_web_page_preview, Boolean
+      attribute :input_message_content, InputMessageContent
+      attribute :reply_markup, InlineKeyboardMarkup
       attribute :url, String
       attribute :hide_url, Boolean
       attribute :description, String
       attribute :thumb_url, String
       attribute :thumb_width, Integer
+      attribute :thumb_height, Integer
     end
   end
 end