+++ lib/telegrammer/data_types/inline_query_result/inline_query_result_mpeg4_gif.rb
@@ -10,9 +10,8 @@ module Telegrammer
     # @attr [String] thumb_url URL of the static thumbnail (jpeg or gif) for the result
     # @attr [String] title Optional. Title for the result
     # @attr [String] caption Optional. Caption of the MPEG-4 file to be sent, 0-200 characters
-    # @attr [String] message_text Optional. Text of a message to be sent instead of the animation, 1-512 characters
-    # @attr [String] parse_mode Optional. Send “Markdown”, if you want Telegram apps to show bold, italic and inline URLs in your bot's message.
-    # @attr [Boolean] disable_web_page_preview Optional. Disables link previews for links in the sent message
+    # @attr [Telegrammer::DataTypes::InlineKeyboardMarkup] reply_markup Optional. Inline keyboard attached to the message
+    # @attr [Telegrammer::DataTypes::InputMessageContent] input_message_content Content of the message to be sent instead of the video animation
     #
     # See more at https://core.telegram.org/bots/api#inlinequeryresultmpeg4gif
     class InlineQueryResultMpeg4Gif < Telegrammer::DataTypes::Base
@@ -24,9 +23,8 @@ module Telegrammer
       attribute :thumb_url, String
       attribute :title, String
       attribute :caption, String
-      attribute :message_text, String
-      attribute :parse_mode, String
-      attribute :disable_web_page_preview, Boolean
+      attribute :reply_markup, InlineKeyboardMarkup
+      attribute :input_message_content, InputMessageContent
     end
   end
 end